+++ src/schemathesis/runner/serialization.py
@@ -16,12 +16,18 @@ from ..utils import format_exception
 class SerializedCase:
     text_lines: List[str] = attr.ib()  # pragma: no mutate
     requests_code: str = attr.ib()
+    path_template: str = attr.ib()
+    path_parameters: Optional[Dict[str, Any]] = attr.ib()
+    query: Optional[Dict[str, Any]] = attr.ib()
 
     @classmethod
     def from_case(cls, case: Case, headers: Optional[Dict[str, Any]]) -> "SerializedCase":
         return cls(
             text_lines=case.as_text_lines(),
             requests_code=case.get_code_to_reproduce(headers),
+            path_template=case.full_path,
+            path_parameters=case.path_parameters,
+            query=case.query,
         )